Release/v0.7.0

+++ lib/src/views/comments/components/dropdown_options_reply.dart
@@ -1,4 +1,5 @@
 import 'package:collection/collection.dart';
+import 'package:feed_sx/src/views/feed/components/post/post_dialog.dart';
 import 'package:likeminds_feed/likeminds_feed.dart';
 import 'package:feed_sx/feed.dart';
 import 'package:feed_sx/src/services/likeminds_service.dart';
@@ -40,27 +41,51 @@ class DropdownOptionsReply extends StatelessWidget {
                   child: Text(element.title),
                   onTap: () async {
                     if (element.title.split(' ').first == "Delete") {
-                      final res =
-                          await locator<LikeMindsService>().getMemberState();
-                      //Implement delete post analytics tracking
-                      LMAnalytics.get().track(
-                        AnalyticsKeys.replyDeleted,
-                        {
-                          "post_id": postId,
-                          "comment_id": commentId,
-                          "reply_id": replyDetails.id,
-                        },
-                      );
-                      final response =
-                          await locator<LikeMindsService>().deleteComment(
-                        DeleteCommentRequest(
-                          postId: postId,
-                          commentId: replyDetails.id,
-                          reason: "Reason for deletion",
-                        ),
-                      );
-                      print(response.toString());
-                      refresh();
+                      showDialog(
+                          context: context,
+                          builder: (childContext) => confirmationDialog(
+                                  childContext,
+                                  title: 'Delete Comment',
+                                  content:
+                                      'Are you sure you want to delete this comment. This action can not be reversed.',
+                                  action: () async {
+                                Navigator.of(childContext).pop();
+                                final res = await locator<LikeMindsService>()
+                                    .getMemberState();
+                                //Implement delete post analytics tracking
+                                LMAnalytics.get().track(
+                                  AnalyticsKeys.replyDeleted,
+                                  {
+                                    "post_id": postId,
+                                    "comment_id": commentId,
+                                    "reply_id": replyDetails.id,
+                                  },
+                                );
+                                final response =
+                                    await locator<LikeMindsService>()
+                                        .deleteComment(
+                                  DeleteCommentRequest(
+                                    postId: postId,
+                                    commentId: replyDetails.id,
+                                    reason: "Reason for deletion",
+                                  ),
+                                );
+                                print(response.toString());
+
+                                if (response.success) {
+                                  ScaffoldMessenger.of(context).showSnackBar(
+                                      confirmationToast(
+                                          content: 'Comment Deleted',
+                                          width: 200,
+                                          backgroundColor: kGrey1Color));
+                                  refresh();
+                                } else {
+                                  ScaffoldMessenger.of(context).showSnackBar(
+                                      confirmationToast(
+                                          content: response.errorMessage ?? '',
+                                          backgroundColor: kGrey1Color));
+                                }
+                              }, actionText: 'Delete'));
                     } else if (element.title.split(' ').first == "Pin") {
                       print("Pinning functionality");
                     }
